allow guest registrar to periodically update guest group
This is helpful when a guest connects to the wifi for the first time. In this case, it is possible for HA appdaemon app to receive the message to add the device to the guest group BEFORE HA has registered the device as an entity. The periodic update will ensure the device is still added to the guests group.